Add Name
Name 클래스를 추가해서 자동차 name의 validate기능을 넘김
