+++ src/main/java/racingcar/domain/Car.java
@@ -1,33 +1,17 @@
 package racingcar.domain;
 
 public class Car {
-    private String name;
+    private Name name;
     private int position;
 
     public Car(String name) {
-        if (isBlank(name)) {
-            throw new IllegalArgumentException("이름 형식이 잘못 되었습니다.");
-        }
-        
-        if (!isSupportFormat(name)) {
-            throw new IllegalArgumentException("이름은 5글자를 초과할 수 없습니다.");
-        }
-
-        this.name = name;
+        this.name = new Name(name);
         this.position = 0;
     }
 
     public Car(String name, int position) {
-        this.name = name;
+        this.name = new Name(name);
         this.position = position;
-    }
-
-    private boolean isSupportFormat(String name) {
-        return name.length() <= 5;
-    }
-
-    private boolean isBlank(String name) {
-        return name == null || name.trim().isEmpty();
     }
 
     public void move(MovingStrategy movingStrategy) {
